Create CustomTableModel
